# academia-front/src/app/services/models/article.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { Domain } from '../models/domain';
import { Feedback } from '../models/feedback';
export interface Article {
  abstract_?: string;
  affiliation?: string;
  approved?: boolean;
  articleCover?: string;
  authorName?: string;
  createdBy?: string;
  createdDate?: string;
  domain?: Domain;
  feedbacks?: Array<Feedback>;
  filePath?: string;
  id?: number;
  isbn?: string;
  lastModifiedBy?: string;
  lastModifiedDate?: string;
  title?: string;
}