# academia-front/src/app/services/models/domain.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { Article } from '../models/article';
export interface Domain {
  articles?: Array<Article>;
  id?: number;
  name?: string;
}