# academia-front/src/openapi/openapi.json
{"openapi":"3.0.1","info":{"title":"OpenApi specification - Academia","description":"OpenApi documentation for Spring Security","termsOfService":"Terms of service","contact":{"name":"Academia","url":"https://academia.com","email":"contact@academia.com"},"license":{"name":"Academia Licence","url":"https://academia-licence.com"},"version":"1.0"},"servers":[{"url":"http://localhost:8088/api/v1","description":"Local ENV"},{"url":"http://academia.com","description":"PROD ENV"}],"security":[{"bearerAuth":[]}],"paths":{"/domains":{"get":{"tags":["domain-controller"],"operationId":"getAllDomains","responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Domain"}}}}}}},"post":{"tags":["domain-controller"],"operationId":"createDomain","parameters":[{"name":"arg0","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Domain"}}}}}}},"/articles":{"get":{"tags":["article-controller"],"operationId":"getAllArticles","responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Article"}}}}}}},"post":{"tags":["article-controller"],"operationId":"createArticle","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Article"}}},"required":true},"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Article"}}}}}}},"/articles/file/{article-id}":{"post":{"tags":["article-controller"],"operationId":"uploadArticleCover","parameters":[{"name":"article-id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"multipart/form-data":{"schema":{"required":["file"],"type":"object","properties":{"file":{"type":"string","format":"binary"}}}}}},"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"object"}}}}}}},"/feedbacks/article/{articleId}":{"get":{"tags":["feedback-controller"],"operationId":"getFeedbacksByArticleId","parameters":[{"name":"articleId","in":"path","required":true,"schema":{"type":"integer","format":"int32"}},{"name":"page","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":0}},{"name":"size","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":10}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PageFeedback"}}}}}}},"/domains/{domainId}/articles":{"get":{"tags":["domain-controller"],"operationId":"getArticlesByDomain","parameters":[],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Article"}}}}}}}},"/articles/{id}":{"get":{"tags":["article-controller"],"operationId":"getArticleById","parameters":[],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Article"}}}}}},"delete":{"tags":["article-controller"],"operationId":"deleteArticle","parameters":[],"responses":{"200":{"description":"OK"}}}},"/articles/my":{"get":{"tags":["article-controller"],"operationId":"getMyArticles","responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Article"}}}}}}}}},"components":{"schemas":{"Article":{"type":"object","properties":{"id":{"type":"integer","format":"int32"},"createdDate":{"type":"string","format":"date-time"},"lastModifiedDate":{"type":"string","format":"date-time"},"createdBy":{"type":"string"},"lastModifiedBy":{"type":"string"},"title":{"type":"string"},"abstract_":{"type":"string"},"isbn":{"type":"string"},"authorName":{"type":"string"},"articleCover":{"type":"string"},"affiliation":{"type":"string"},"filePath":{"type":"string"},"approved":{"type":"boolean"},"feedbacks":{"type":"array","items":{"$ref":"#/components/schemas/Feedback"}},"domain":{"$ref":"#/components/schemas/Domain"}}},"Domain":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"type":"string"},"articles":{"type":"array","items":{"$ref":"#/components/schemas/Article"}}}},"Feedback":{"type":"object","properties":{"id":{"type":"integer","format":"int32"},"createdDate":{"type":"string","format":"date-time"},"lastModifiedDate":{"type":"string","format":"date-time"},"createdBy":{"type":"string"},"lastModifiedBy":{"type":"string"},"note":{"type":"number","format":"double"},"comment":{"type":"string"},"article":{"$ref":"#/components/schemas/Article"}}},"PageFeedback":{"type":"object","properties":{"totalPages":{"type":"integer","format":"int32"},"totalElements":{"type":"integer","format":"int64"},"pageable":{"$ref":"#/components/schemas/PageableObject"},"size":{"type":"integer","format":"int32"},"content":{"type":"array","items":{"$ref":"#/components/schemas/Feedback"}},"number":{"type":"integer","format":"int32"},"sort":{"$ref":"#/components/schemas/SortObject"},"first":{"type":"boolean"},"last":{"type":"boolean"},"numberOfElements":{"type":"integer","format":"int32"},"empty":{"type":"boolean"}}},"PageableObject":{"type":"object","properties":{"pageNumber":{"type":"integer","format":"int32"},"pageSize":{"type":"integer","format":"int32"},"offset":{"type":"integer","format":"int64"},"sort":{"$ref":"#/components/schemas/SortObject"},"paged":{"type":"boolean"},"unpaged":{"type":"boolean"}}},"SortObject":{"type":"object","properties":{"sorted":{"type":"boolean"},"empty":{"type":"boolean"},"unsorted":{"type":"boolean"}}}},"securitySchemes":{"bearerAuth":{"type":"oauth2","description":"JWT auth description","in":"header","scheme":"bearer","bearerFormat":"JWT","flows":{"clientCredentials":{"authorizationUrl":"http://localhost:9090/realms/book-social-network/protocol/openid-connect/auth","scopes":{}}}}}}}